feat: add-friends-likes 09 (#2)
* feat: add-friends-likes 09

* refactor: add-friends-likes 09

* #2 refactor: add-friends-likes 09

+++ src/main/java/ru/yandex/practicum/filmorate/service/IdGenerator.java
@@ -5,14 +5,14 @@ import org.springframework.stereotype.Service;
 @Service
 public class IdGenerator {
 
-    private volatile int filmCounter = 1;
-    private volatile int userCounter = 1;
+    private volatile long filmCounter = 1;
+    private volatile long userCounter = 1;
 
-    public synchronized int generateFilmId() {
+    public synchronized long generateFilmId() {
         return filmCounter++;
     }
 
-    public synchronized int generateUserId() {
+    public synchronized long generateUserId() {
         return userCounter++;
     }
